app/requirements/base.txt
Django==4.2.1
djangorestframework==3.14.0
django-cors-headers==4.0.0
drf_spectacular==0.26.2
factory_boy==3.2.1
coverage==6.4.4
black==23.3.0
flake8==6.0.0
isort==5.12.0
django-debug-toolbar==4.1.0
requests